update and test
