+++ app/components/BackButton.tsx
@@ -1,35 +1,39 @@
-import React from 'react';
-import { ArrowLeft } from 'lucide-react';
+"use client";
+import React from "react";
+import { svgs } from "./svgs";
 
-
-interface BackButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  variant: "expired" | "createCapsule";
-  showIcon?: boolean;
-  text?: string;
+interface BackButtonProps {
+  variant: "plain" | "iconed";
+  buttonAction: () => void;
 }
 
-const BackButton = React.forwardRef<HTMLButtonElement, BackButtonProps>(
-  ({ variant, showIcon = true, text = "Back", className = '', ...props }, ref) => {
-    return (
-      <div className="flex flex-col items-center"> {/* Container to place label below */}
-      <div >   
-
-       
-      <button
-          ref={ref}
-          className={w-16 h-16 bg-emerald-500 rounded-full flex items-center justify-center shadow-md ${className}}
-          aria-label={text}
-          type="button"
-          {...props}
-        >
-           <ArrowLeft className="w-10 h-10 text-white" /> 
-        </button>
+const BackButton: React.FC<BackButtonProps> = ({ variant, buttonAction }) => {
+  return variant === "plain" ? (
+    <button
+      onClick={buttonAction}
+      className="flex items-center justify-center w-32 gap-2"
+    >
+      <svgs.BackIcon2 />
+      <p className="font-kumbhSans text-base font-normal text-[#111111] text-nowrap border-b-2 border-[#111111]">
+        Back
+      </p>
+    </button>
+  ) : (
+    <button
+      onClick={buttonAction}
+      className="flex flex-col items-start justify-center w-32"
+    >
+      <svgs.BackIcon />
+      <div className="flex items-center justify-start ml-4">
+        <p className="font-kumbhSans text-base font-medium text-[#48BB78] text-nowrap border-b-2 border-[#48BB78]">
+          Go Back
+        </p>
+        <span className="translate-x-[20px] translate-y-[30px]">
+          <svgs.BackArrow />
+        </span>
       </div>
-        <span className="mt-2 text-xl font-medium text-green-500 underline decoration-2 ">{text}</span> 
-      </div>
-    );
-  }
-);
+    </button>
+  );
+};
 
-BackButton.displayName = 'BackButton';
 export default BackButton;
